--- a/src/rust/engine/src/externs/interface.rs
+++ b/src/rust/engine/src/externs/interface.rs
@@ -79,6 +79,7 @@ fn native_engine(py: Python, m: &Bound<'_, PyModule>) -> PyO3Result<()> {
     externs::workunits::register(m)?;
     externs::dep_inference::register(m)?;
     externs::unions::register(py, m)?;
+    externs::frozendict::register(py, m)?;
 
     m.add("PollTimeout", py.get_type::<PollTimeout>())?;