… src/python/pants/engine/internals/native_engine.pyi