Before:
```console
(pants@3.11) ➜  pants git:(839363c8c9) hyperfine -w 2 -r 10 './pants --no-dynamic-ui --no-pantsd dependencies ::'
Benchmark 1: ./pants --no-dynamic-ui --no-pantsd dependencies ::
  Time (mean ± σ):     21.010 s ±  0.201 s    [User: 21.579 s, System: 7.112 s]
  Range (min … max):   20.804 s … 21.366 s    10 runs
 ```
After:
```console
(pants@3.11) ➜  pants git:(add/port-frozen-dict) hyperfine -w 2 -r 10 './pants --no-dynamic-ui --no-pantsd dependencies ::'
Benchmark 1: ./pants --no-dynamic-ui --no-pantsd dependencies ::
  Time (mean ± σ):     20.436 s ±  0.095 s    [User: 20.932 s, System: 7.101 s]
  Range (min … max):   20.332 s … 20.605 s    10 runs
```

~2.5% faster :man_shrugging: