@tdyas 2025-07-15
Does [PEP 0603](https://peps.python.org/pep-0603/) ("Adding a frozenmap type to collections") have any relevance? I.e., could we just re-use a frozen mapping type already on PyPI (especially if it uses a native module for speed)? For example, https://github.com/MagicStack/immutables is cited in the PEP as just such a type.