@tobni 2025-07-15
@tdyas Hadn't considered it but likely to provide similar or even better performance benefits? I'm not sure whether creating glue for a thirdparty package is more or less ergonomic than having control of the collections ourselves however.