@tobni 2025-07-15
Migrating types to rust makes it easier to write intrinsic rules. I'd like to see e.g target api:s be rust code.