@benjyw 2025-07-28
Code looks good so far although I still want to dive deeper. But I'm curious why there is a perf improvement? It looks like the underlying data structures are still the Python ones.